requirements.txt
numpy~=1.18.5
pandas~=1.1.5
matplotlib~=3.5.1
scikit-learn~=1.0.2
tqdm~=4.62.3
ase~=3.22.1
quippy-ase~=0.9.9
